requirements.txt
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
graphviz==0.20.3
imageio==2.34.2
kiwisolver==1.4.5
matplotlib==3.9.1.post1
neat-python==0.92
networkx==3.3
numpy==2.0.1
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.2
pillow==10.4.0
pygame==2.6.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
visualize==0.5.1
